#bugfix - x64 version of dir_list failing per email from Paul 12/26/14 - trace to type casting issue and add sandbox test project. lk_stdlib opendir and rewinddir and DIR struct updated.
git-svn-id: https://efmsvn2.nrel.gov/lk/svn/trunk@123 beec71b5-fff0-e442-8ac2-4f84a11f714e

+++ lk_stdlib.cpp
@@ -2657,8 +2657,9 @@ bool lk::tex_doc( const lk_string &file,
 
 struct DIR
 {
-    long                handle; /* -1 for failed rewind */
-    struct _finddata_t  info;
+	intptr_t                handle; /* -1 for failed rewind */
+	//long                handle; /* -1 for failed rewind */
+	struct _finddata_t  info;
     struct dirent       result; /* d_name null iff first time */
     char                *name;  /* null-terminated char string */
 };
@@ -2678,8 +2679,9 @@ DIR *opendir(const char *name)
         {
             strcat(strcpy(dir->name, name), all);
 
-            if((dir->handle = (long) _findfirst(dir->name, &dir->info)) != -1)
-            {
+			if ((dir->handle = _findfirst(dir->name, &dir->info)) != -1)
+//				if ((dir->handle = (long)_findfirst(dir->name, &dir->info)) != -1)
+				{
                 dir->result.d_name = 0;
             }
             else /* rollback */
@@ -2698,7 +2700,7 @@ DIR *opendir(const char *name)
     }
     else
     {
-        errno = EINVAL;
+        errno = EINVAL; 
     }
 
     return dir;
@@ -2752,8 +2754,9 @@ void rewinddir(DIR *dir)
     if(dir && dir->handle != -1)
     {
         _findclose(dir->handle);
-        dir->handle = (long) _findfirst(dir->name, &dir->info);
-        dir->result.d_name = 0;
+//		dir->handle = (long)_findfirst(dir->name, &dir->info);
+		dir->handle = _findfirst(dir->name, &dir->info);
+		dir->result.d_name = 0;
     }
     else
     {
